requirements.txt

---


```txt
torch
torchvision
transformers
ultralytics
opencv-python
numpy
Pillow
scikit-learn
matplotlib
pyyaml
psutil
